njila-reporting-service/requirements.txt
anyio==4.12.1
asgiref==3.11.1
async-timeout==5.0.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==5.2.12
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
dnspython==2.8.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.3
pillow==12.1.1
psycopg2-binary==2.9.11
py_eureka_client==0.13.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
redis==7.3.0
reportlab==4.4.10
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
